Add human readable timestamp
This commit presents a human readable date format for message timestamps.

Why?

* Unfortunately, the default timestamp `%Y:%m:%dT%H:%M:%S.xxZ` is not
human readable.

Changes:

* Users can get the timestamp in more readable form of `%d %b %Y, %H:%M:%S`.
As an example, `03 Feb 2020, 18:30:00`

Affected Classes:

* `config` file gets a new member under `news` keyword, namely
timestamp. It represents the user's local timezone in the UTC format. As
an example, Turkey's local timezone is UTC +3.00, so the timezone field
should be "3.00".

* `newsArticle` gets a new function, `getHumanReadableDate`, accepting
raw timestamp and local timezone. The local timestamp is parsed to the
form of list of two integers, they represent hours and minutes
respectively.

* `newsReader` also gets a new member, called timezone. Since we always
create `newsArticle`, performing the parsing operation in the
constructor made a lot of sense, to avoid unnecessary computations.

This commit should resolve the issue #10.